vishal/requirements.txt
groq==0.4.2
python-dotenv==1.0.1